requirements.txt
discord.py==2.6.3
pipx==1.7.1
pipenv==2024.4.0
python-dotenv
requests
libnspr4==latest
libxcb-shm0==latest
libxcomposite1==latest
libxdamage1==latest
libxfixes3==latest
libxcb1==latest
libgtk-3-0t64==latest
libpangocairo-1.0-0==latest
libpango-1.0-0==latest
libatk1.0-0t64==latest
libcairo-gobject2==latest
libcairo2==latest
libgdk-pixbuf-2.0-0==latest
libxrender1==latest
libasound2t64==latest
libdbus-1-3==latest
